summon commands listed in core
Bootstrap will only contain very few profiles by default from now on.
The goal is to not have these be in the way custom profiles.
The necessary commands to download these profiles have been added to the 'core' profile in the 'install tools' cell.
The original profiles have been split into sets and are hosted on dedicated repos.